test(uhost): gate live lifecycle test
